refactor: extract the dapp entity into its own feature domain
