+++ src/components/DappBrowser/Homepage.tsx
@@ -30,11 +30,12 @@ import { IS_TEST } from '@/env';
 import { FEATURED_RESULTS } from '@/features/config/constants/experimental';
 import { useExperimentalFlag } from '@/features/config/hooks/experimentalHooks';
 import { useRemoteConfig } from '@/features/config/stores/remoteConfig';
+import { useTrendingDApps } from '@/features/dapp/resources/trendingDapps';
+import { getDappHost } from '@/features/dapp/utils/dappUrls';
 import { HTTP_PREFIX, HTTPS_PREFIX } from '@/framework/core/utils/url';
 import { opacity } from '@/framework/ui/utils/opacity';
 import { type DApp } from '@/graphql/__generated__/metadata';
 import * as i18n from '@/languages';
-import { useTrendingDApps } from '@/resources/metadata/trendingDapps';
 import { useBrowserStore } from '@/state/browser/browserStore';
 import { useFavoriteDappsStore, type FavoritedSite } from '@/state/browser/favoriteDappsStore';
 import { useBrowserHistoryStore, type Site } from '@/state/browserHistory';
@@ -46,7 +47,6 @@ import { EasingGradient } from '../easing-gradient/EasingGradient';
 import { useBrowserContext } from './BrowserContext';
 import { HOMEPAGE_BACKGROUND_COLOR_DARK, HOMEPAGE_BACKGROUND_COLOR_LIGHT } from './constants';
 import { EXTRA_WEBVIEW_HEIGHT, WEBVIEW_HEIGHT } from './Dimensions';
-import { getDappHost } from './handleProviderRequest';
 import { getNameFromFormattedUrl } from './utils';
 
 const HORIZONTAL_PAGE_INSET = 24;
