+++ src/features/dapp/resources/trendingDapps.ts
