+++ src/features/wallet-connect/screens/AuthRequest.tsx
@@ -9,11 +9,11 @@ import ButtonPressAnimation from '@/components/animations/ButtonPressAnimation';
 import { ImgixImage } from '@/components/images';
 import { InfoAlert } from '@/components/info-alert/info-alert';
 import { AccentColorProvider, BackgroundProvider, Box, Separator, Text } from '@/design-system';
+import { useDappMetadata } from '@/features/dapp/resources/dapp';
 import { DAppStatus } from '@/graphql/__generated__/metadata';
 import * as i18n from '@/languages';
 import { useNavigation } from '@/navigation/Navigation';
 import Routes from '@/navigation/routesNames';
-import { useDappMetadata } from '@/resources/metadata/dapp';
 import { getAccountProfileInfo, useAccountAddress, useSelectedWallet } from '@/state/wallets/walletsStore';
 import { useTheme } from '@/theme/ThemeContext';
 import { initials } from '@/utils/formatters';
